+++ LLama/Extensions/IModelParamsExtensions.cs
@@ -12,28 +12,6 @@ namespace LLama.Extensions
     public static class IModelParamsExtensions
     {
         /// <summary>
-        /// Convert the given `IModelParams` into a `LLamaContextParams`
-        /// </summary>
-        /// <param name="params"></param>
-        /// <param name="result"></param>
-        /// <returns></returns>
-        /// <exception cref="FileNotFoundException"></exception>
-        /// <exception cref="ArgumentException"></exception>
-        public static void ToLlamaContextParams(this IContextParams @params, out LLamaContextParams result)
-        {
-            result = NativeApi.llama_context_default_params();
-            result.n_ctx = @params.ContextSize;
-            result.n_batch = @params.BatchSize;
-            result.seed = @params.Seed;
-            result.f16_kv = @params.UseFp16Memory;
-            result.logits_all = @params.Perplexity;
-            result.embedding = @params.EmbeddingMode;
-            result.rope_freq_base = @params.RopeFrequencyBase;
-            result.rope_freq_scale = @params.RopeFrequencyScale;
-            result.mul_mat_q = @params.MulMatQ;
-        }
-
-        /// <summary>
         /// Convert the given `IModelParams` into a `LLamaModelParams`
         /// </summary>
         /// <param name="params"></param>
